Refactor dashboard into components with CSS modules
