+++ src/components/AppLayout/AppLayout.module.css
@@ -1,0 +1,22 @@
+.layout {
+  min-height: 100vh;
+  background: #f5f6fa;
+}
+
+.header {
+  background: #fff;
+  padding: 24px 32px;
+  border-bottom: 1px solid #e5e6eb;
+}
+
+.content {
+  margin: 24px;
+}
+
+.title {
+  margin: 0 !important;
+}
+
+.subtitle {
+  color: rgba(0, 0, 0, 0.45);
+}
