+++ src/components/MetricsSummary/MetricsSummary.module.css
@@ -1,0 +1,6 @@
+.grid {
+  display: grid;
+  gap: 16px;
+  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+  margin-bottom: 24px;
+}
