+++ src/data/mockFunds.ts
@@ -1,0 +1,140 @@
+import type { FundProduct } from "../types/fund";
+
+export const mockFunds: FundProduct[] = [
+  {
+    id: "F001",
+    fundName: "聚势多资产进取一号",
+    ticker: "JSMA-01",
+    manager: "刘强",
+    strategy: "多资产",
+    inceptionDate: "2019-06-30",
+    navStart: 1.28,
+    navEnd: 1.46,
+    stockExposureStart: 52,
+    stockExposureEnd: 58,
+    futuresExposureStart: 12,
+    futuresExposureEnd: 15,
+    aum: 2.6e9,
+    sharpe: 1.62,
+    ytdReturn: 0.118,
+  },
+  {
+    id: "F002",
+    fundName: "稳健量化中性基金",
+    ticker: "QHZX-11",
+    manager: "张伟",
+    strategy: "量化对冲",
+    inceptionDate: "2021-01-15",
+    navStart: 1.12,
+    navEnd: 1.19,
+    stockExposureStart: 8,
+    stockExposureEnd: 6,
+    futuresExposureStart: 35,
+    futuresExposureEnd: 38,
+    aum: 3.1e9,
+    sharpe: 1.87,
+    ytdReturn: 0.072,
+  },
+  {
+    id: "F003",
+    fundName: "卓越固收增强",
+    ticker: "GUGS-07",
+    manager: "王芳",
+    strategy: "固收+",
+    inceptionDate: "2018-09-01",
+    navStart: 1.09,
+    navEnd: 1.21,
+    stockExposureStart: 18,
+    stockExposureEnd: 20,
+    futuresExposureStart: 5,
+    futuresExposureEnd: 4,
+    aum: 4.5e9,
+    sharpe: 1.45,
+    ytdReturn: 0.089,
+  },
+  {
+    id: "F004",
+    fundName: "精选行业增强",
+    ticker: "ZXZQ-28",
+    manager: "陈晨",
+    strategy: "指数增强",
+    inceptionDate: "2020-04-12",
+    navStart: 0.98,
+    navEnd: 1.18,
+    stockExposureStart: 82,
+    stockExposureEnd: 88,
+    futuresExposureStart: 10,
+    futuresExposureEnd: 9,
+    aum: 1.9e9,
+    sharpe: 1.32,
+    ytdReturn: 0.153,
+  },
+  {
+    id: "F005",
+    fundName: "灵动套利一号",
+    ticker: "LDTL-18",
+    manager: "周航",
+    strategy: "量化对冲",
+    inceptionDate: "2022-02-18",
+    navStart: 1.03,
+    navEnd: 1.11,
+    stockExposureStart: 14,
+    stockExposureEnd: 17,
+    futuresExposureStart: 42,
+    futuresExposureEnd: 39,
+    aum: 1.4e9,
+    sharpe: 1.74,
+    ytdReturn: 0.064,
+  },
+  {
+    id: "F006",
+    fundName: "优享稳健回报",
+    ticker: "YXWJ-05",
+    manager: "李娜",
+    strategy: "固收+",
+    inceptionDate: "2017-11-20",
+    navStart: 1.26,
+    navEnd: 1.35,
+    stockExposureStart: 12,
+    stockExposureEnd: 13,
+    futuresExposureStart: 0,
+    futuresExposureEnd: 1,
+    aum: 5.2e9,
+    sharpe: 1.28,
+    ytdReturn: 0.058,
+  },
+  {
+    id: "F007",
+    fundName: "高端制造主题增强",
+    ticker: "GZZT-03",
+    manager: "赵峰",
+    strategy: "指数增强",
+    inceptionDate: "2021-07-09",
+    navStart: 1.14,
+    navEnd: 1.33,
+    stockExposureStart: 76,
+    stockExposureEnd: 83,
+    futuresExposureStart: 16,
+    futuresExposureEnd: 14,
+    aum: 2.2e9,
+    sharpe: 1.51,
+    ytdReturn: 0.132,
+  },
+  {
+    id: "F008",
+    fundName: "全球配置旗舰",
+    ticker: "QQPZ-02",
+    manager: "孙宇",
+    strategy: "多资产",
+    inceptionDate: "2016-03-30",
+    navStart: 1.42,
+    navEnd: 1.58,
+    stockExposureStart: 48,
+    stockExposureEnd: 51,
+    futuresExposureStart: 22,
+    futuresExposureEnd: 25,
+    aum: 6.8e9,
+    sharpe: 1.67,
+    ytdReturn: 0.096,
+  },
+];
