+++ src/utils/formatters.ts
@@ -1,0 +1,11 @@
+export const formatterCurrency = new Intl.NumberFormat("zh-CN", {
+  style: "currency",
+  currency: "CNY",
+  maximumFractionDigits: 0,
+});
+
+export const formatterPercent = new Intl.NumberFormat("zh-CN", {
+  style: "percent",
+  minimumFractionDigits: 2,
+  maximumFractionDigits: 2,
+});
